Add service.io module
